requirements.txt
gspread
google-auth
google-auth-oauthlib
google-cloud-storage
requests
beautifulsoup4
functions-framework
